+++ manager/allocator/cnmallocator/portallocator_test.go
@@ -1,4 +1,4 @@
-package networkallocator
+package cnmallocator
 
 import (
 	"testing"
