--- manager/allocator/networkallocator/drivers_unsupported.go
@@ -1,10 +1,0 @@
-// +build !linux,!darwin,!windows
-
-package networkallocator
-
-const initializers = nil
-
-// PredefinedNetworks returns the list of predefined network structures
-func PredefinedNetworks() []PredefinedNetworkData {
-	return nil
-}
